+++ NEWS.md
@@ -1,7 +1,5 @@
 # phinterval (development version)
 
-# phinterval 1.0.0.9000
-
 ## Features:
 
 * New `phint_cumunion()`, `phint_cumintersect()`, `phint_symmetric_setdiff()` to complete the family of set-operations.
@@ -10,7 +8,7 @@
   
   * `phint_symmetric_setdiff()` takes the symmetric set-difference of two phintervals.
   
-* New `phint_flatten()` returns all spans or gaps within an entire phinterval vector as a vector of intervals.
+* New `phint_flatten()` and `datetime_flatten()` return all spans or gaps within an entire phinterval vector or datetime spans as a vector of intervals.
 
 ## Bug fixes: